+++ src/commands/signal.rs
@@ -33,7 +33,7 @@ impl SignalArgs {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test_util::test_utils;
+    extern crate test_utils;
     use anyhow::Context;
     use std::{
         env,