+++ tests/cli.rs
@@ -10,12 +10,10 @@ use std::{
     thread,
 };
 
-mod util;
-
 #[test]
 #[serial]
 fn list_command_no_apps_running() -> Result<()> {
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.arg("list");
     cmd.assert()
         .success()
@@ -25,7 +23,7 @@ fn list_command_no_apps_running() -> Result<()> {
 
 #[test]
 fn log_no_apps_running() -> Result<()> {
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["log", "test_app_not_available"]);
     cmd.assert()
         .failure()
@@ -35,7 +33,7 @@ fn log_no_apps_running() -> Result<()> {
 
 #[test]
 fn send_no_apps_running() -> Result<()> {
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["send", "test_app_not_available", "command"]);
     cmd.assert()
         .failure()
@@ -45,7 +43,7 @@ fn send_no_apps_running() -> Result<()> {
 
 #[test]
 fn signal_no_apps_running() -> Result<()> {
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["signal", "test_app_not_available", "0"]);
     cmd.assert()
         .failure()
@@ -55,7 +53,7 @@ fn signal_no_apps_running() -> Result<()> {
 
 #[test]
 fn status_no_apps_running() -> Result<()> {
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["status", "test_app_not_available"]);
     cmd.assert()
         .failure()
@@ -65,7 +63,7 @@ fn status_no_apps_running() -> Result<()> {
 
 #[test]
 fn stop_no_apps_running() -> Result<()> {
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["stop", "test_app_not_available", "-f"]);
     cmd.assert()
         .failure()
@@ -75,7 +73,7 @@ fn stop_no_apps_running() -> Result<()> {
 
 #[test]
 fn kill_no_apps_running() -> Result<()> {
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["kill", "test_app_not_available"]);
     cmd.assert()
         .failure()
@@ -85,7 +83,7 @@ fn kill_no_apps_running() -> Result<()> {
 
 #[test]
 fn attach_no_apps_running() -> Result<()> {
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["attach", "test_app_not_available"]);
     cmd.assert()
         .failure()
@@ -96,17 +94,17 @@ fn attach_no_apps_running() -> Result<()> {
 #[test]
 fn stop_force_long_running_service() -> Result<()> {
     let name = "stop_force_long_running_service";
-    util::start_long_running_service(name)?;
-    assert!(util::check_app_is_running(name)?);
+    test_utils::start_long_running_service(name)?;
+    assert!(test_utils::check_app_is_running(name)?);
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["stop", name, "-f"]);
 
     cmd.assert()
         .success()
         .stdout(predicate::str::contains("Signal sent."));
 
-    util::delete_app_folder(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
@@ -114,13 +112,13 @@ fn stop_force_long_running_service() -> Result<()> {
 fn stop_long_running_service() -> Result<()> {
     let name = "stop_long_running_service";
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["start", "-n", name, "-p", "example"]);
     cmd.assert().success();
 
-    assert!(util::check_app_is_running(name)?);
+    assert!(test_utils::check_app_is_running(name)?);
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["stop", name]);
 
     cmd.assert()
@@ -130,25 +128,32 @@ fn stop_long_running_service() -> Result<()> {
     // Sleeping to make sure the process shutdown
     thread::sleep(std::time::Duration::from_secs(1));
 
-    util::delete_app_folder(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
 #[test]
 fn profile_example_profile() -> Result<()> {
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["profile", "example"]);
     cmd.assert()
         .success()
         .stdout(predicate::str::contains("tools"));
+
+    cmd = test_utils::get_base_command();
+    cmd.args(["profile", "example", "--json"]);
+    cmd.assert()
+        .success()
+        .stdout(predicate::str::contains("\"interpreter_arguments\""));
+
     Ok(())
 }
 
 #[test]
 fn start_short_lived_command() -> Result<()> {
     let name = "start_echo";
-    util::start_short_lived_command(name)?;
-    util::delete_app_folder(name)?;
+    test_utils::start_short_lived_command(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
@@ -156,7 +161,7 @@ fn start_short_lived_command() -> Result<()> {
 fn start_python_without_interpreter() -> Result<()> {
     let name = "start_python_without_interpreter";
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args([
         "start",
         "./tools/long_running_service.py",
@@ -187,53 +192,53 @@ fn start_python_without_interpreter() -> Result<()> {
     let pids: Vec<&str> = pid_str.lines().collect();
     assert!(pids.len() == 1);
 
-    util::delete_app_folder(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
 #[test]
 fn log_short_lived_command() -> Result<()> {
     let name = "log_echo";
-    util::start_short_lived_command(name)?;
+    test_utils::start_short_lived_command(name)?;
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["log", name]);
 
     cmd.assert()
         .success()
         .stdout(predicate::str::contains(">> Printed"));
 
-    util::delete_app_folder(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
 #[test]
 fn log_empty_file() -> Result<()> {
     let name = "log_empty_file";
-    util::start_short_lived_command(name)?;
+    test_utils::start_short_lived_command(name)?;
 
     let home = env::var("HOME").context("Error getting HOME env.")?;
     let log_dir = PathBuf::from(home).join(".crescent/apps/log_empty_file/log_empty_file.log");
     fs::remove_file(&log_dir)?;
     File::create(log_dir)?;
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["log", name]);
 
     cmd.assert()
         .success()
         .stdout(predicate::str::contains("Log is empty at the moment."));
 
-    util::delete_app_folder(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
 #[test]
 fn log_follow_short_lived_command() -> Result<()> {
     let name = "log_follow_echo";
-    util::start_short_lived_command(name)?;
+    test_utils::start_short_lived_command(name)?;
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
 
     cmd.args(["log", name, "-f"])
         .timeout(std::time::Duration::from_secs(1));
@@ -251,23 +256,23 @@ fn log_follow_short_lived_command() -> Result<()> {
         Err(err) => return Err(anyhow!("{err}")),
     };
 
-    util::delete_app_folder(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
 #[test]
 fn log_flush_command() -> Result<()> {
     let name = "log_flush_command";
-    util::start_short_lived_command(name)?;
+    test_utils::start_short_lived_command(name)?;
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["log", name, "--flush"]);
 
     cmd.assert().success().stdout(predicate::str::contains(
         "Flushed 'log_flush_command' log file.",
     ));
 
-    util::delete_app_folder(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
@@ -275,25 +280,57 @@ fn log_flush_command() -> Result<()> {
 #[serial]
 fn list_command_long_running_service() -> Result<()> {
     let name = "list_long_running_service";
-    util::start_long_running_service(name)?;
-    assert!(util::check_app_is_running(name)?);
+    test_utils::start_long_running_service(name)?;
+    assert!(test_utils::check_app_is_running(name)?);
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["list"]);
 
     cmd.assert()
         .success()
         .stdout(predicate::str::contains(name));
 
-    util::shutdown_long_running_service(name)?;
-    util::delete_app_folder(name)?;
+    test_utils::shutdown_long_running_service(name)?;
+    test_utils::delete_app_folder(name)?;
+    Ok(())
+}
+
+#[test]
+#[serial]
+fn start_saved_long_running_service() -> Result<()> {
+    let name = "start_saved_running_service";
+    test_utils::start_long_running_service(name)?;
+    assert!(test_utils::check_app_is_running(name)?);
+
+    let mut cmd = test_utils::get_base_command();
+    cmd.args(["save"]);
+
+    cmd.assert()
+        .success()
+        .stdout(predicate::str::contains("Saved "));
+
+    test_utils::shutdown_long_running_service(name)?;
+
+    cmd = test_utils::get_base_command();
+    cmd.args(["start", "--saved"]);
+
+    cmd.assert()
+        .success()
+        .stdout(predicate::str::contains("Starting applications."));
+
+    // Sleeping to make sure the process started
+    thread::sleep(std::time::Duration::from_secs(1));
+
+    assert!(test_utils::check_app_is_running(name)?);
+    test_utils::shutdown_long_running_service(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
 #[test]
 fn start_long_running_service_with_profile() -> Result<()> {
     let name = "example";
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["start", "-p", "example"]);
 
     cmd.assert()
@@ -303,9 +340,9 @@ fn start_long_running_service_with_profile() -> Result<()> {
     // Sleeping to make sure the process started
     thread::sleep(std::time::Duration::from_secs(1));
 
-    assert!(util::check_app_is_running(name)?);
+    assert!(test_utils::check_app_is_running(name)?);
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["status", name]);
 
     let binding = cmd.assert().success();
@@ -321,60 +358,60 @@ fn start_long_running_service_with_profile() -> Result<()> {
         Err(err) => return Err(anyhow!("{err}")),
     };
 
-    util::shutdown_long_running_service(name)?;
-    util::delete_app_folder(name)?;
+    test_utils::shutdown_long_running_service(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
 #[test]
 fn signal_long_running_service() -> Result<()> {
     let name = "signal_long_running_service";
-    util::start_long_running_service(name)?;
-    assert!(util::check_app_is_running(name)?);
+    test_utils::start_long_running_service(name)?;
+    assert!(test_utils::check_app_is_running(name)?);
 
-    util::shutdown_long_running_service(name)?;
-    util::delete_app_folder(name)?;
+    test_utils::shutdown_long_running_service(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
 #[test]
 fn send_command_socket() -> Result<()> {
     let name = "send_socket_test";
-    util::start_long_running_service(name)?;
-    assert!(util::check_app_is_running(name)?);
+    test_utils::start_long_running_service(name)?;
+    assert!(test_utils::check_app_is_running(name)?);
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["send", name, "ping"]);
 
     cmd.assert()
         .success()
         .stdout(predicate::str::contains("Command sent."));
 
-    util::shutdown_long_running_service(name)?;
-    util::delete_app_folder(name)?;
+    test_utils::shutdown_long_running_service(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
 #[test]
 fn attach_short_lived_command() -> Result<()> {
     let name = "attach_echo";
-    util::start_short_lived_command(name)?;
+    test_utils::start_short_lived_command(name)?;
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["attach", name]);
 
     cmd.assert()
         .failure()
         .stderr(predicate::str::contains("Application not running."));
 
-    util::delete_app_folder(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }
 
 #[test]
 fn attach_command_socket_not_found() -> Result<()> {
     let name = "attach_socket_not_found";
-    util::start_short_lived_command(name)?;
+    test_utils::start_short_lived_command(name)?;
 
     let home = env::var("HOME").context("Error getting HOME env.")?;
     let socket_dir = PathBuf::from(home)
@@ -384,13 +421,13 @@ fn attach_command_socket_not_found() -> Result<()> {
         fs::remove_file(socket_dir)?
     }
 
-    let mut cmd = util::get_base_command();
+    let mut cmd = test_utils::get_base_command();
     cmd.args(["attach", name]);
 
     cmd.assert()
         .failure()
         .stderr(predicate::str::contains("Application not running."));
 
-    util::delete_app_folder(name)?;
+    test_utils::delete_app_folder(name)?;
     Ok(())
 }